requirements.txt
asgiref==3.8.1
Django==5.1.5
django-nested-admin==4.1.1
python-monkey-business==1.1.0
sqlparse==0.5.3
